bind events on player after it's ready
This gives the player time to do it's own work before the plugin handles the event.

There is a case where on the `fullscreenchange` event videojs toggles the `vjs-fullscreen` class. If the plugin handler runs before that, the values of ima.height() is wrong (is 0) when player is using the intrinsic ratio class (`.vjs-16-9`) for it's dimensions.