fix : h5Player.src uses "blob:http" scheme
we cannot compare this.h5Player.src with this.contentSrc if we use https://github.com/streamroot/videojs-hlsjs-plugin (dailyMotion hlsjs engine), because this.h5Player.src uses blob-http or blob-https prefix scheme (not an-ordinary http or https prefix scheme). https://superuser.com/questions/948738/what-is-the-blobhttp-prefix-and-where-can-i-learn-more-about-this
so we must avoid reguler step for make this procedure fixed.